First draft of historical price data

+++ harmony/rpc/smartContract.go
@@ -26,7 +26,7 @@ func SimpleCall(to, data string) (r string, err error) {
 }
 
 // HistoricCall executes a read only transaction to the specified contract with given data at the given block.
-func HistoricCall(to, data string, blockNum int) (r string, err error) {
+func HistoricCall(to, data string, blockNum uint64) (r string, err error) {
 	// Get return value
 	params := []interface{}{
 		struct {
